week02_management_and_testing/homework/pyproject.toml
[project]
name = "effdl-homework-01"
version = "0.1.0"
description = "lol really"
readme = "README.md"
requires-python = ">=3.10"

dependencies = [
    "torch==2.1.2",
    "torchvision==0.16.2",
    "tqdm==4.66.1",
    "numpy==1.26.4",
    "wandb==0.13.10",
    "dvc==2.44.0",
    "hydra-core==1.3.1",
    "omegaconf==2.3.0",
]

[tool.uv]
dev-dependencies = [
    "pytest==7.4.4",
    "pytest-cov==4.1.0",
]
